requirements.txt
# AI Job Agent — Mac M2 Optimized
# Core pipeline dependencies

# ─── Web Automation ────────────────────────────────────────────────────────────
playwright>=1.40.0

# ─── LLM / Ollama ─────────────────────────────────────────────────────────────
requests>=2.31.0
ollama>=0.1.7

# ─── Embeddings & Similarity ───────────────────────────────────────────────────
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4            # CPU version — perfect for Mac M2
numpy>=1.24.0

# ─── API & Dashboard ──────────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# ─── Data & Storage ───────────────────────────────────────────────────────────
# sqlite3 is built into Python — no install needed

# ─── Scheduler & Utils ────────────────────────────────────────────────────────
schedule>=1.2.0
psutil>=5.9.7               # CPU/RAM monitoring

# ─── QLoRA Fine-tuning (optional — only if fine-tuning locally) ───────────────
# Uncomment these if you want to run QLoRA fine-tuning:
# torch>=2.1.0
# transformers>=4.36.0
# peft>=0.7.0
# datasets>=2.16.0
# trl>=0.7.0
# accelerate>=0.25.0
# bitsandbytes>=0.41.0      # Note: limited support on Mac — use mlx-lm instead

# ─── Alternative for Mac M2 Fine-tuning (Apple MLX) ─────────────────────────
# pip install mlx-lm          # Native Apple Silicon fine-tuning
